\name{gaussian}
\alias{gaussian}
\alias{N80}
\alias{N128}
\alias{N160}
\alias{N256}
\alias{N512}
\alias{N640}
\docType{data}
\title{
  Tables for (reduced) gaussian grids from ECMWF
}
\description{ 
These tables contain the latitudes and the number of longitudes for various gaussian grid resolutions from ECMWF. Part of the effort for reading global data.
}
\keyword{datasets}
