\name{gribtables}
\alias{specialnames}
\alias{extratab}
\alias{extratab2}
\docType{data}
\title{
  GRIB table2 versions that may not (yet) be recognised by grib_api.
}
\description{\itemize{ 
\item \code{extratab} contains grib tables not available in grib_api
\item \code{extratab2} contains grib tables not available in grib_api
\item \code{specialnames} contains some simple rules to get similar 'shortName' results as grib_api in special cases.
}}
\keyword{datasets}
